Update lift.py
I updated the way that this has named the dataframe, as even on a copy the inplace edits caused issues since both were named data.  I also added in the ability to take a full model.  It probably needs some testing in terms of model arguments and the transformation function, and currently doesn't work with categorical (which I am fine with).